move detail func + fix paddings
