+++ demehin.maxim/T1/creating_utils.hpp
@@ -1,0 +1,14 @@
+#ifndef CREATING_UTILS_HPP
+#define CREATING_UTILS_HPP
+#include <cstddef>
+#include "base-types.hpp"
+
+namespace demehin
+{
+  namespace details
+  {
+    point_t* copyCords(const point_t* vertex, size_t vrt_cnt);
+  }
+}
+
+#endif
