+++ demehin.maxim/T1/factory.cpp
@@ -3,10 +3,10 @@
 #include "rectangle.hpp"
 #include "ring.hpp"
 #include "polygon.hpp"
+#include "creating_utils.hpp"
 
 namespace
 {
-
   void inputParameters(std::istream& in, double* cords, size_t cords_cnt)
   {
     for (size_t i = 0; i < cords_cnt; i++)
@@ -19,7 +19,6 @@ namespace
   {
     double rectCords[4] = {};
     inputParameters(in, rectCords, 4);
-
     demehin::point_t left_bot = {rectCords[0], rectCords[1]};
     demehin::point_t right_top = {rectCords[2], rectCords[3]};
 
@@ -30,7 +29,6 @@ namespace
   {
     double ringParams[4] = {};
     inputParameters(in, ringParams, 4);
-
     demehin::point_t center = {ringParams[0], ringParams[1]};
 
     return new demehin::Ring(center, ringParams[2], ringParams[3]);
@@ -41,7 +39,6 @@ namespace
     demehin::point_t* pts = new demehin::point_t[capacity];
     size_t size = 0;
     double vrt_x = 0, vrt_y = 0;
-
     while (in >> vrt_x >> vrt_y)
     {
       if (size >= capacity)
@@ -60,10 +57,12 @@ namespace
         delete[] pts;
         pts = new_pts;
       }
+
       pts[size++] = {vrt_x, vrt_y};
       cord_cnt++;
     }
     in.clear();
+
     return pts;
   }
 
@@ -83,27 +82,26 @@ namespace
       throw;
     }
     delete[] vrt;
+
     return plg;
   }
 }
 
-namespace demehin
+void demehin::createShape(std::istream& in, const std::string& shp_name, Shape** shapes, size_t& shp_cnt)
 {
-  void createShape(std::istream& in, const std::string& shp_name, Shape** shapes, size_t& shp_cnt)
+  if (shp_name == "RECTANGLE")
   {
-    if (shp_name == "RECTANGLE")
-    {
-      shapes[shp_cnt++] = createRect(in);
-    }
+    shapes[shp_cnt++] = createRect(in);
+  }
 
-    else if (shp_name == "RING")
-    {
-      shapes[shp_cnt++] = createRing(in);
-    }
+  else if (shp_name == "RING")
+  {
+    shapes[shp_cnt++] = createRing(in);
+  }
 
-    else if (shp_name == "POLYGON")
-    {
+  else if (shp_name == "POLYGON")
+  {
       shapes[shp_cnt++] = createPolygon(in);
-    }
   }
 }
+
