+++ demehin.maxim/T1/rectangle.hpp
@@ -13,7 +13,7 @@ namespace demehin
     rectangle_t getFrameRect() const override;
     void move(point_t s) override;
     void move(double x, double y) override;
-    void unsafeScale(double k) override;
+    void scaleUnsafely(double k) override;
   private:
     Polygon plg_;
   };
