+++ demehin.maxim/T1/shape.hpp
@@ -12,7 +12,7 @@ namespace demehin
     virtual rectangle_t getFrameRect() const = 0;
     virtual void move(point_t s) = 0;
     virtual void move(double x, double y) = 0;
-    virtual void unsafeScale(double k) = 0;
+    virtual void scaleUnsafely(double k) = 0;
     virtual void scale(double k);
   };
 }
